deleteFolder이 폴더존재여부 검증

+++ src/main/java/maestrogroup/core/folder/FolderService.java
@@ -39,6 +39,9 @@ public class FolderService {
     }
 
     public void deleteFolder(int folderIdx) throws BaseException{
+        if(folderDao.isExistsFolder(folderIdx) != 1){
+            throw new BaseException(BaseResponseStatus.NOT_EXISTS_FOLDER);
+        }
         //유효한 folderIdx인지에 대한 검증
         //User가 Folder를 삭제하고자 하는 팀에 속해있는지에 대한 검증
         folderDao.deleteFolder(folderIdx);
